+++ NvimView/NvimView/CocoaExtensions.swift
@@ -48,3 +48,18 @@ extension NSView {
     return Array(UnsafeBufferPointer(start: rectsPtr, count: count))
   }
 }
+
+extension NSEvent.ModifierFlags {
+
+  // Values are from https://github.com/SFML/SFML/blob/master/src/SFML/Window/OSX/SFKeyboardModifiersHelper.mm
+  // @formatter:off
+  static let rightShift   = NSEvent.ModifierFlags(rawValue: 0x020004)
+  static let leftShift    = NSEvent.ModifierFlags(rawValue: 0x020002)
+  static let rightCommand = NSEvent.ModifierFlags(rawValue: 0x100010)
+  static let leftCommand  = NSEvent.ModifierFlags(rawValue: 0x100008)
+  static let rightOption  = NSEvent.ModifierFlags(rawValue: 0x080040)
+  static let leftOption   = NSEvent.ModifierFlags(rawValue: 0x080020)
+  static let rightControl = NSEvent.ModifierFlags(rawValue: 0x042000)
+  static let leftControl  = NSEvent.ModifierFlags(rawValue: 0x040001)
+  // @formatter:on
+}